modularize login, character select, and game screens
The class for each screen accepts `main` as the first parameter to the
constructor, and the `Main` class keeps a reference to each screen as well as
to the `ConnectionManager`. This way, all classes can obtain to a reference to
any of the other classes using `main` as a sort of hub.