Move model and animation logic from Player class into Character class
By doing this, we can reuse `Character` for networked players as well.
Unfortunately, it is somewhat hackish since we most likely catch user input
faster than the network tickrate, and so every time we want to move the local
player, we overwrite that character's animation tickrate to be based on `dt`.

Also includes clamping on camera distance.