improve chat interface
We used to error out if Game.camera was not yet initialized but the user tried
to send a message. Ideally this shouldn't ever happen but covering these edge
cases will be very helpful.

Additionally, whispering to an otherwise empty server would still enter the
chat UI. Now we just display the error and stay out of chat mode.

This also gets rid of the debug line for displaying outgoing messages locally.